scraper/requirements.txt
python-dotenv==1.0.1
selenium==4.28.1
supabase==2.13.0
webdriver_manager==4.0.2
